Add ID columns to book and order list tables
In `BackOffice.cshtml`, added a new column to display the ID of each book in the book list table. This includes adding `<th>#</th>` in the table header and `<td class="align-middle">@book.Id</td>` in the table body.

Similarly, added a new column to display the ID of each order in the order list table. This includes adding `<th>#</th>` in the table header and `<td class="align-middle">@order.Id</td>` in the table body.